+++ templates/tabs/analyse_account.html
@@ -285,7 +285,7 @@
             </table>
             <div id="result_section_3" style="text-align:center; padding-top: 10px;">
               <a class="btn btn-primary" 
-              href="https://twitter.com/intent/tweet?text=Analysis%20of%20{{result.tweetsetInfo.term | urlencode }}%3A%0ASentiment%3A%20%20%20%20%20%20%09{{result.tweetsetInfo.summary | urlencode }}%0APolitics%3A%20%20%20%20%20%20%20%20%20%09{{result.political_sentiment_data.pol_summary | urlencode }}%0AStrongest%20Emotions%3A%20{% for item in result.most_used_data.emotion_summary %}{{item.name | urlencode }}{% if not loop.last %},&nbsp;{%endif%}{% endfor %}%0AMost%20Used%20Words%3A%20%20%20%20{% for item in result.most_used_data.most_used_words %}{{item.word | urlencode }}{% if not loop.last %},&nbsp;{%endif%}{% endfor %}{% if result.most_used_data.most_used_emojis != [] %}%0AMost%20Used%20Emojis%3A%20%20%20{% for item in result.most_used_data.most_used_emojis %}{{item.word | urlencode }}{% if not loop.last %},&nbsp;{% else %}&#010;{%endif%}{% endfor %}{% endif %}%0AAnalysed%20with%3A%20%09%20%20%23TwitterAnalysisApplication" 
+              href="https://twitter.com/intent/tweet?text=Analysis%20of%20{{result.tweetsetInfo.term | urlencode }}%3A%0ASentiment%3A%20%20%20%20%20%20%09{{result.tweetsetInfo.summary | urlencode }}%0APolitics%3A%20%20%20%20%20%20%20%20%20%09{{result.political_sentiment_data.pol_summary | urlencode }}%0AStrongest%20Emotions%3A%20{% for item in result.most_used_data.emotion_summary %}{{item | urlencode }}{% if not loop.last %},&nbsp;{%endif%}{% endfor %}%0AMost%20Used%20Words%3A%20%20%20%20{% for item in result.most_used_data.most_used_words %}{{item.word | urlencode }}{% if not loop.last %},&nbsp;{%endif%}{% endfor %}{% if result.most_used_data.most_used_emojis != [] %}%0AMost%20Used%20Emojis%3A%20%20%20{% for item in result.most_used_data.most_used_emojis %}{{item.word | urlencode }}{% if not loop.last %},&nbsp;{% else %}&#010;{%endif%}{% endfor %}{% endif %}%0AAnalysed%20with%3A%20%09%20%20%23TwitterAnalysisApplication" 
               target="_blank">
                 <img src="{{ url_for('static',filename='images/twitter_logo.png') }}" class="verified_tick"/> 
                 Share on Twitter
